% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\docType{data}
\name{Formatter}
\alias{Formatter}
\title{Stub formatter}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Formatter
}
\arguments{
\item{entry}{entry should have, at minimum, fields message, timestamp, level}
}
\description{
Stub formatter
}
\keyword{datasets}
