% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{JSONFormatter}
\alias{JSONFormatter}
\title{json formatter factory function}
\usage{
JSONFormatter(auto_unbox = TRUE, field_map = NULL, ...)
}
\arguments{
\item{auto_unbox}{pass to jsonlite}

\item{field_map}{rename internal fields not exposed to user}

\item{...}{additional fields passed to jsonlite and glue}
}
\description{
json formatter factory function
}
\details{
TODO: add details
}
