% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outputs.R
\docType{data}
\name{LogOutput}
\alias{LogOutput}
\title{LogOutput base class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
LogOutput
}
\description{
LogOutput base class
}
\keyword{datasets}
