% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{TextFormatter}
\alias{TextFormatter}
\title{text formatter factory function}
\usage{
TextFormatter(format_string = "[{level}] {message} {extras}",
  no_color = FALSE, field_map = NULL)
}
\arguments{
\item{format_string}{glue format string}

\item{no_color}{suppress colored output}

\item{field_map}{rename internal fields not exposed to user}

\item{...}{additional fields passed to glue}
}
\description{
text formatter factory function
}
\details{
TODO: add details
}
