% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{JSONFormatter}
\alias{JSONFormatter}
\title{json formatter factory function}
\usage{
JSONFormatter(field_map = NULL, .auto_unbox = TRUE, .force = TRUE,
  ...)
}
\arguments{
\item{field_map}{rename internal fields not exposed to user}

\item{.auto_unbox}{pass to jsonlite}

\item{.force}{force as much coercion as possible for json conversion}

\item{...}{additional fields passed to jsonlite and glue}
}
\description{
json formatter factory function
}
\details{
TODO: add details
}
