% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\docType{data}
\name{TextFormatter}
\alias{TextFormatter}
\title{text formatter}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TextFormatter
}
\description{
text formatter
}
\keyword{datasets}
