% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters.R
\name{TextFormatter}
\alias{TextFormatter}
\title{text formatter factory function}
\usage{
TextFormatter(format_string = "{level}: {message} {extras}",
  field_map = NULL, no_color = FALSE, no_truncate = FALSE,
  no_space = FALSE, ...)
}
\arguments{
\item{format_string}{glue format string}

\item{field_map}{rename internal fields not exposed to user}

\item{no_color}{suppress colored output}

\item{no_truncate}{stop truncation of levels to 4 characters}

\item{no_space}{no additional spacing for message fields}

\item{...}{additional fields passed to glue}
}
\description{
text formatter factory function
}
\details{
TODO: add details
}
